2D & 3D vectors done
Finished the 2D & 3D vector classes and requirements for the classes.
Documentation has also been completed. New error code system has also
been created and implemented into the library. Time to move onto the 4D
vector class and the last of the simple math operators. After that it's
time for unit testing and work on the Wikipage.